+++ packages/pangea-node-sdk/tests/unit/utils.test.ts
@@ -1,7 +1,14 @@
 import { it, expect } from "@jest/globals";
-import crc32c from "@node-rs/crc32";
-import * as fs from "fs";
-import { hashSHA1, hashSHA256, getHashPrefix, strToB64, b64toStr, hashNTLM } from "@src/index.js";
+import fs from "fs";
+import {
+  hashSHA1,
+  hashSHA256,
+  getHashPrefix,
+  strToB64,
+  b64toStr,
+  hashNTLM,
+  hashCRC32C,
+} from "@src/index.js";
 import { getFileUploadParams } from "@src/utils/utils.js";
 
 const testfilePath = "./tests/testdata/testfile.pdf";
@@ -30,19 +37,18 @@ it("Base 64 strings", async () => {
 
 it("hashNTLM test", async () => {
   const msg = "password";
-  const hash = hashNTLM(msg);
+  const hash = await hashNTLM(msg);
   expect(hash).toBe("8846F7EAEE8FB117AD06BDD830B7586C");
 });
 
 it("CRC32C test", async () => {
   const msg = "ABCDEF";
-  let crcValue = crc32c.crc32c(msg, 0) >>> 0;
-  const crc = crcValue.toString(16);
+  let crc = await hashCRC32C(Buffer.from(msg));
   expect(crc).toBe("a4b7ce68");
 });
 
 it("getFileParams test filepath", async () => {
-  const paramsFilepath = getFileUploadParams(testfilePath);
+  const paramsFilepath = await getFileUploadParams(testfilePath);
   expect(paramsFilepath.crc32c).toBe("754995fb");
   expect(paramsFilepath.sha256).toBe(
     "81655950d560e804a6315e09e74a7414e7b18ba99f722abe6122857e69a3aebd"
@@ -52,7 +58,7 @@ it("getFileParams test filepath", async () => {
 
 it("getFileParams test buffer", async () => {
   const file = fs.readFileSync(testfilePath);
-  const paramsFilepath = getFileUploadParams(file);
+  const paramsFilepath = await getFileUploadParams(file);
 
   expect(paramsFilepath.crc32c).toBe("754995fb");
   expect(paramsFilepath.sha256).toBe(